requirements.txt
python-dotenv==1.0.0
pandas
numpy
scikit-learn
shap
xgboost
google-generative-ai
